stuff - some styling

+++ config.go
@@ -17,9 +17,9 @@ type ConfigEnv struct {
 
 	DefaultExpiryTime   string
 	NoFileUpload        bool
-	MaxFileSize         int64
-	MaxEncryptionSize   int64
-	MaxContentLength    int64
+	MaxFileSize         int
+	MaxEncryptionSize   int
+	MaxContentLength    int
 	UploadingPassword   string
 	DisableEternalPaste bool
 	DisableReadCount    bool
@@ -44,9 +44,9 @@ func InitConfig() {
 		PublicList:          getEnv("PUBLIC_LIST", "true") == "true",
 		PublicURL:           getEnv("PUBLIC_URL", "http://localhost:"+getEnv("PORT", ":9454")),
 		NoFileUpload:        getEnv("NO_FILE_UPLOAD", "false") == "true",
-		MaxFileSize:         int64(getEnvInt("MAX_FILE_SIZE", "1024 * 1024 * 10")),
-		MaxEncryptionSize:   int64(getEnvInt("MAX_ENCRYPTION_SIZE", "1024 * 1024 * 10")),
-		MaxContentLength:    int64(getEnvInt("MAX_CONTENT_LENGTH", "1024 * 1024")),
+		MaxFileSize:         getEnvInt("MAX_FILE_SIZE", "1024 * 1024 * 10"),
+		MaxEncryptionSize:   getEnvInt("MAX_ENCRYPTION_SIZE", "1024 * 1024 * 10"),
+		MaxContentLength:    getEnvInt("MAX_CONTENT_LENGTH", "1024 * 1024"),
 		UploadingPassword:   getEnv("UPLOADING_PASSWORD", ""),
 		DisableEternalPaste: getEnv("DISABLE_ETERNAL_PASTE", "false") == "true",
 		DisableReadCount:    getEnv("DISABLE_READ_COUNT", "false") == "true",
